refactor: PBI-10-View API Monitor Detail-Frontend (#40)
* [REFACTOR] give the better color for various data

* [RED] change the input test for the better chart implementation

* [GREEN] move process data logic of each chart to each chart file

* [REFACTOR] rename the variable and classname to be more representative
